Rename sync response and break down large file
