+++ Packages/EurofurenceKit/Sources/EurofurenceKit/Remote/Responses/RemoteEntity.swift
@@ -1,0 +1,1 @@
+protocol RemoteEntity: Decodable, Identifiable where Self.ID == String { }
