+++ Packages/EurofurenceKit/Sources/EurofurenceKit/Remote/Responses/RemoteLink.swift
@@ -1,0 +1,13 @@
+struct RemoteLink: Decodable {
+    
+    private enum CodingKeys: String, CodingKey {
+        case fragmentType = "FragmentType"
+        case name = "Name"
+        case target = "Target"
+    }
+    
+    var fragmentType: String
+    var name: String
+    var target: String
+    
+}
