+++ Packages/EurofurenceKit/Sources/EurofurenceKit/Remote/Responses/RemoteTrack.swift
@@ -1,0 +1,15 @@
+import Foundation.NSDate
+
+struct RemoteTrack: RemoteEntity {
+    
+    private enum CodingKeys: String, CodingKey {
+        case lastChangeDateTimeUtc = "LastChangeDateTimeUtc"
+        case id = "Id"
+        case name = "Name"
+    }
+    
+    var lastChangeDateTimeUtc: Date
+    var id: String
+    var name: String
+    
+}
